perf : join using jpa entity, repository
