+++ src/main/java/org/example/shoppingweather/controller/CustomerApiController.java
@@ -2,6 +2,7 @@ package org.example.shoppingweather.controller;
 
 
 import lombok.RequiredArgsConstructor;
+import org.example.shoppingweather.dto.AddUserRequest;
 import org.example.shoppingweather.dto.SignUpRequestDTO;
 import org.example.shoppingweather.dto.SignUpResponseDTO;
 import org.example.shoppingweather.service.CustomerService;
@@ -9,9 +10,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping("/user")
 @RequiredArgsConstructor
 public class CustomerApiController {
 
@@ -19,11 +22,11 @@ public class CustomerApiController {
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
     @PostMapping("/join")
-    public ResponseEntity<SignUpResponseDTO> signUp(@RequestBody SignUpRequestDTO signUpRequestDTO) {
-        customerService.signUp(signUpRequestDTO.toCustomer(bCryptPasswordEncoder));
+    public ResponseEntity<SignUpResponseDTO> signup(@RequestBody SignUpRequestDTO signUpRequestDTO) {
+        customerService.save(signUpRequestDTO); // 회원가입 메소드 호출
         return ResponseEntity.ok(
                 SignUpResponseDTO.builder()
-                        .url("/user/login")
+                        .url("/user/login") // 회원 가입이 완료된 후 로그인 페이지로 이동
                         .build()
         );
     }
