--- src/main/java/org/example/shoppingweather/controller/CustomerController.java
@@ -1,21 +1,0 @@
-package org.example.shoppingweather.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/user")
-public class CustomerController {
-
-    @GetMapping("/join")
-    public String signUp() {
-        return "join";
-    }
-
-    @GetMapping("/login")
-    public String signIn() {
-        return "login";
-    }
-
-}
