--- src/main/java/org/example/shoppingweather/controller/UserApiController.java
@@ -1,23 +1,0 @@
-package org.example.shoppingweather.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.example.shoppingweather.dto.AddUserRequest;
-import org.example.shoppingweather.service.UserService;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequiredArgsConstructor
-@RequestMapping("/user")
-public class UserApiController {
-
-    private final UserService userService;
-
-    @PostMapping("/join")
-    public String signup(AddUserRequest request) {
-        userService.save(request); // 회원가입 메소드 호출
-        return "redirect:/login"; // 회원 가입이 완료된 후 로그인 페이지로 이동
-    }
-
-}
