+++ src/main/java/org/example/shoppingweather/repository/CustomerRepository.java
@@ -9,4 +9,5 @@ import java.util.Optional;
 @Repository
 public interface CustomerRepository extends JpaRepository<Customer, Long> {
     Optional<Customer> findByLoginId(String loginId);
+
 }
