+++ src/main/resources/templates/join.html
@@ -9,7 +9,6 @@
     <link th:href="@{/css/heather.css}" rel="stylesheet">
     <link th:href="@{/css/footer.css}" rel="stylesheet">
 
-
     <title>회원가입 페이지</title>
 </head>
 <body>
